EHUB-147: Backend code optimized

+++ application/src/main/java/com/entando/hub/catalog/rest/BundleGroupController.java
@@ -1,22 +1,8 @@
 package com.entando.hub.catalog.rest;
 
-import com.entando.hub.catalog.persistence.entity.Organisation;
-import com.entando.hub.catalog.rest.BundleGroupVersionController.BundleGroupVersionView;
-import com.entando.hub.catalog.service.BundleGroupService;
-import com.entando.hub.catalog.service.BundleGroupVersionService;
-import com.entando.hub.catalog.service.CategoryService;
-import com.entando.hub.catalog.service.security.SecurityHelperService;
-import io.swagger.v3.oas.annotations.Operation;
-import lombok.*;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.data.domain.Page;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import javax.annotation.security.RolesAllowed;
-import javax.transaction.Transactional;
+import static com.entando.hub.catalog.config.AuthoritiesConstants.ADMIN;
+import static com.entando.hub.catalog.config.AuthoritiesConstants.AUTHOR;
+import static com.entando.hub.catalog.config.AuthoritiesConstants.MANAGER;
 
 import java.util.Arrays;
 import java.util.List;
@@ -24,7 +10,37 @@ import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-import static com.entando.hub.catalog.config.AuthoritiesConstants.*;
+import javax.annotation.security.RolesAllowed;
+import javax.transaction.Transactional;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.data.domain.Page;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+
+import com.entando.hub.catalog.persistence.entity.Organisation;
+import com.entando.hub.catalog.rest.BundleGroupVersionController.BundleGroupVersionView;
+import com.entando.hub.catalog.service.BundleGroupService;
+import com.entando.hub.catalog.service.BundleGroupVersionService;
+import com.entando.hub.catalog.service.CategoryService;
+import com.entando.hub.catalog.service.security.SecurityHelperService;
+
+import io.swagger.v3.oas.annotations.Operation;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.Setter;
+import lombok.ToString;
 
 @RestController
 @RequestMapping("/api/bundlegroups")
@@ -35,17 +51,14 @@ public class BundleGroupController {
     private final BundleGroupService bundleGroupService;
     private final CategoryService categoryService;
     private final SecurityHelperService securityHelperService;
-    
     private final BundleGroupVersionService bundleGroupVersionService;
 
-    public BundleGroupController(BundleGroupService bundleGroupService, CategoryService categoryService, SecurityHelperService securityHelperService,
-    		BundleGroupVersionService bundleGroupVersionService) {
+    public BundleGroupController(BundleGroupService bundleGroupService, CategoryService categoryService, SecurityHelperService securityHelperService, BundleGroupVersionService bundleGroupVersionService) {
         this.bundleGroupService = bundleGroupService;
         this.categoryService = categoryService;
         this.securityHelperService = securityHelperService;
         this.bundleGroupVersionService = bundleGroupVersionService;
     }
-
 
     //PUBLIC
     @Operation(summary = "Get all the bundle groups in the hub", description = "Public api, no authentication required. You can provide the organisationId.")
@@ -55,7 +68,6 @@ public class BundleGroupController {
         logger.debug("REST request to get BundleGroups by organisation Id: {}", organisationId);
         return bundleGroupService.getBundleGroups(Optional.ofNullable(organisationId)).stream().map(BundleGroup::new).collect(Collectors.toList());
     }
-
 
     //PUBLIC
     @Operation(summary = "Get all the bundle groups in the hub", description = "Public api, no authentication required. You can provide the organisationId the categoryIds and the statuses [NOT_PUBLISHED, PUBLISHED, PUBLISH_REQ, DELETE_REQ, DELETED]")
@@ -94,7 +106,6 @@ public class BundleGroupController {
             return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
         }
     }
-
 
     @Operation(summary = "Create a new bundleGroup", description = "Protected api, only eh-admin, eh-author or eh-manager can access it.")
     @RolesAllowed({ADMIN, AUTHOR, MANAGER})
@@ -151,18 +162,17 @@ public class BundleGroupController {
     public ResponseEntity<CategoryController.Category> deleteBundleGroup(@PathVariable String bundleGroupId) {
         logger.debug("REST request to delete bundleGroup {}", bundleGroupId);
         Optional<com.entando.hub.catalog.persistence.entity.BundleGroup> bundleGroupOptional = bundleGroupService.getBundleGroup(bundleGroupId);
-//        if (!bundleGroupOptional.isPresent() || !bundleGroupOptional.get().getStatus().equals(com.entando.hub.catalog.persistence.entity.BundleGroupVersion.Status.DELETE_REQ)) {
-//            bundleGroupOptional.ifPresentOrElse(
-//                    bundleGroup -> logger.warn("Requested bundleGroup '{}' is not in DELETE_REQ status: {}", bundleGroupId, bundleGroup.getStatus()),
-//                    () -> logger.warn("Requested bundleGroup '{}' does not exists", bundleGroupId)
-//            );
-//            return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
-//        } else {
+        if (!bundleGroupOptional.isPresent()) {
+            bundleGroupOptional.ifPresentOrElse(
+                    bundleGroup -> logger.warn("Requested bundleGroup '{}' is not present", bundleGroupId),
+                    () -> logger.warn("Requested bundleGroup '{}' does not exists", bundleGroupId)
+            );
+            return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
+        } else {
             bundleGroupService.deleteBundleGroup(bundleGroupId);
             return new ResponseEntity<>(null, HttpStatus.NO_CONTENT);
-//        }
+        }
     }
-
 
     @Getter
     @Setter
@@ -185,14 +195,6 @@ public class BundleGroupController {
     @Data
     public static class BundleGroupNoId {
         protected final String name;
-//        protected final String description;
-//        protected final String descriptionImage;
-//        protected String documentationUrl;
-//        protected String bundleGroupUrl;
-//        protected String version;
-//        protected com.entando.hub.catalog.persistence.entity.BundleGroup.Status status;
-//        protected LocalDateTime lastUpdate;
-
         //the following must be merged with the entity using mappedBy
         protected List<String> children;
         protected String organisationId;
@@ -200,36 +202,18 @@ public class BundleGroupController {
         protected List<String> categories;
         protected BundleGroupVersionView versionDetails;
 
-
-
         public BundleGroupNoId(String name ,String organisationId) {
             this.name = name;
             this.organisationId = organisationId;
-//            this.description = description;
-//            this.descriptionImage = descriptionImage;
-//            this.version = version;
         }
 
-
         public BundleGroupNoId(com.entando.hub.catalog.persistence.entity.BundleGroup entity) {
-//            this.description = entity.getDescription();
-//            this.descriptionImage = entity.getDescriptionImage();
-              this.name = entity.getName();
-//            this.status = entity.getStatus();
-//            this.documentationUrl = entity.getDocumentationUrl();
-//            this.bundleGroupUrl = entity.getBundleGroupUrl();
-//            this.version = entity.getVersion();
-//            this.lastUpdate = entity.getLastUpdate();
+            this.name = entity.getName();
 
             if (entity.getOrganisation() != null) {
                 this.organisationId = entity.getOrganisation().getId().toString();
                 this.organisationName = entity.getOrganisation().getName();
             }
-            //todo one single iteration
-//            kamlesh: commented
-//            if (entity.getBundles() != null) {
-//                this.children = entity.getBundles().stream().map((children) -> children.getId().toString()).collect(Collectors.toList());
-//            }
             if (entity.getCategories() != null) {
                 this.categories = entity.getCategories().stream().map((category) -> category.getId().toString()).collect(Collectors.toList());
             }
@@ -237,12 +221,7 @@ public class BundleGroupController {
 
         public com.entando.hub.catalog.persistence.entity.BundleGroup createEntity(Optional<String> id) {
             com.entando.hub.catalog.persistence.entity.BundleGroup ret = new com.entando.hub.catalog.persistence.entity.BundleGroup();
-//            ret.setDescription(this.getDescription());
-              ret.setName(this.getName());
-//            ret.setDescriptionImage(this.getDescriptionImage());
-//            ret.setDocumentationUrl(this.getDocumentationUrl());
-//            ret.setStatus(this.getStatus());
-//            ret.setVersion(this.getVersion());
+            ret.setName(this.getName());
             if (this.organisationId != null) {
                 Organisation organisation = new Organisation();
                 organisation.setId(Long.parseLong(this.organisationId));
@@ -253,6 +232,4 @@ public class BundleGroupController {
             return ret;
         }
     }
-
-
 }
